tcti_device: Move convenience functions to common directory.
Both the InitDeviceTctiContext and TeardownDeviceTcti functions are not
part of the SAPI/TCTI spec. This commit moves them from the exposed
interface for the device TCTI to common/tcti_device_util.[ch]. AFAIK
these functions were created for convenience in the resourcemgr and test
applications.
