+++ tcti/tcti_device.c
@@ -48,8 +48,6 @@
 
 #define HOSTNAME_LENGTH 200
 
-const char *deviceTctiName = "device TCTI";
-
 int (*tpmLocalTpmPrintf)( printf_type type, const char *format, ...) = DebugPrintf;
 
 TSS2_RC LocalTpmSendTpmCommand(
@@ -275,28 +273,5 @@ TSS2_RC InitDeviceTcti (
         }
     }
 
-    return rval;
-}
-
-TSS2_RC TeardownDeviceTcti(TSS2_TCTI_CONTEXT *tctiContext)
-{
-    ((TSS2_TCTI_CONTEXT_INTEL *)tctiContext)->finalize( tctiContext );
-
-    return TSS2_RC_SUCCESS;
-}
-
-TSS2_RC InitDeviceTctiContext( const TCTI_DEVICE_CONF *driverConfig, TSS2_TCTI_CONTEXT **tctiContext )
-{
-    size_t size;
-    
-    TSS2_RC rval = TSS2_RC_SUCCESS;
-
-    rval = InitDeviceTcti(NULL, &size, driverConfig, 0, 0, deviceTctiName );
-    if( rval != TSS2_RC_SUCCESS )
-        return rval;
-    
-    *tctiContext = malloc(size);
-
-    rval = InitDeviceTcti(*tctiContext, &size, driverConfig, TCTI_MAGIC, TCTI_VERSION, deviceTctiName );
     return rval;
 }
